Change promoted link to use #find
