+++ lib/outbrain/api/promoted_link.rb
@@ -2,15 +2,11 @@ require 'outbrain/api/report'
 
 module Outbrain
   module Api
-    class PromotedLink < Report
-      PATH="promotedLinks"
+    class PromotedLink < Base
+      PATH = "promotedLinks"
 
-      def self.path(options)
-        "campaigns/#{options.fetch(:campaign_id)}/promotedLinks"
-      end
-
-      def self.resource_name
-        "promotedLinks"
+      def self.find(id)
+          Request.find( PATH, id, { as: self })
       end
     end
   end
