+++ lib/outbrain/api/promoted_link_report.rb
@@ -1,0 +1,21 @@
+require 'outbrain/api/report'
+
+module Outbrain
+  module Api
+    class PromotedLinkReport < Report
+      PATH="promotedLinks"
+
+      def self.path(options)
+        "campaigns/#{options.fetch(:campaign_id)}/promotedLinks"
+      end
+
+      def self.resource_name
+        "promotedLinks"
+      end
+
+      def self.find(id, options)
+          Request.find(PATH, id, { as: self })
+      end
+    end
+  end
+end
